refactoring authorization
