+++ src/styles/_customButton.module.scss
@@ -2,7 +2,7 @@
 @import "variable";
 
 .loginButton {
-  @extend .textHeading;
+  @extend .textFont;
   display: flex;
   height: 52px;
   width: 540px;
